# Conflicts:
#	SpaceTrader/app/src/main/java/com/example/spacetrader/entity/MarketGood.java
#	SpaceTrader/app/src/main/java/com/example/spacetrader/views/CreatePlayerActivity.java

+++ SpaceTrader/app/src/main/java/com/example/spacetrader/views/CreatePlayerActivity.java
@@ -116,7 +116,7 @@ public class CreatePlayerActivity extends AppCompatActivity {
                     int trader = (int) trader_spinner.getSelectedItem();
                     int engineer = (int) engineer_spinner.getSelectedItem();
 
-                    Integer pts_left = (Integer) (16 - pilot - fighter - trader - engineer);
+                    Integer pts_left = (16 - pilot - fighter - trader - engineer);
                     skill_pts.setText(pts_left.toString());
 
                 }
